+++ typescript/packages/plugins/uniswap/package.json
@@ -11,28 +11,14 @@
     "main": "./dist/index.js",
     "module": "./dist/index.mjs",
     "types": "./dist/index.d.ts",
-    "devDependencies": {
-        "@types/node": "^22.7.4",
-        "globals": "^15.12.0",
-        "ts-node": "^10.9.2",
-        "tsup": "^8.3.5",
-        "typescript": "^5.6.3",
-        "vitest": "^2.1.5"
-    },
     "dependencies": {
         "@goat-sdk/core": "workspace:*",
-        "viem": "^2.21.49",
+        "viem": "catalog:",
         "zod": "catalog:"
     },
     "peerDependencies": {
         "@goat-sdk/core": "workspace:*",
-        "viem": "^2.21.49"
-    },
-    "engines": {
-        "node": ">=20.12.2 <21",
-        "npm": "please-use-pnpm",
-        "pnpm": ">=9",
-        "yarn": "please-use-pnpm"
+        "viem": "catalog:"
     },
     "homepage": "https://ohmygoat.dev",
     "repository": {
@@ -43,6 +29,5 @@
     "bugs": {
         "url": "https://github.com/goat-sdk/goat/issues"
     },
-    "keywords": ["ai", "agents", "web3"],
-    "packageManager": "pnpm@9.14.2"
+    "keywords": ["ai", "agents", "web3"]
 }
