+++ typescript/packages/plugins/uniswap/src/types/UniswapCtorParams.ts
@@ -1,0 +1,4 @@
+export type UniswapCtorParams = {
+    apiKey: string;
+    baseUrl: string;
+};
